+++ slinky/src/keep_sections.rs
@@ -5,11 +5,10 @@ use std::collections::HashSet;
 
 use serde::Deserialize;
 
-// use crate::{absent_nullable::AbsentNullable, traits::Serial, Settings, SlinkyError};
-
 #[derive(Debug, Clone, PartialEq, Eq, Default, Deserialize)]
 #[serde(deny_unknown_fields)]
 #[serde(untagged)]
+#[non_exhaustive]
 pub enum KeepSections {
     #[default]
     #[serde(skip)]