+++ slinky/src/linker_symbols_style.rs
@@ -7,6 +7,7 @@ use crate::utils;
 
 #[derive(Deserialize, PartialEq, Debug)]
 #[serde(rename_all = "snake_case")]
+#[non_exhaustive]
 pub enum LinkerSymbolsStyle {
     Splat,
     Makerom,